added models for Departure/PitStop/Destination

+++ db/migrate/20221022151148_create_road_trips.rb
@@ -2,7 +2,7 @@ class CreateRoadTrips < ActiveRecord::Migration[7.0]
   def change
     create_table :road_trips do |t|
       t.string :trip_name
-      t.bigint :user_id
+      t.int :user_id
       t.string :road_trip_distance_miles
       t.integer :user_likes
